requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
Django==5.1.2
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
packaging==24.1
psycopg2-binary==2.9.9
PyJWT==2.9.0
sqlparse==0.5.1
typing_extensions==4.12.2
whitenoise==6.7.0
